requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements-dev.in
#
asgiref==3.5.2
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   django
attrs==22.1.0
    # via pytest
backports-zoneinfo==0.2.1
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   django
cfgv==3.3.1
    # via pre-commit
distlib==0.3.6
    # via virtualenv
django==4.1.4
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==3.13.0
    # via
    #   -c requirements.txt
    #   -r requirements.txt
django-environ==0.9.0
    # via
    #   -c requirements.txt
    #   -r requirements.txt
djangorestframework==3.14.0
    # via
    #   -c requirements.txt
    #   -r requirements.txt
exceptiongroup==1.0.4
    # via pytest
factory-boy==3.2.0
    # via
    #   -c requirements.txt
    #   -r requirements.txt
faker==1.0.7
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   factory-boy
filelock==3.8.2
    # via virtualenv
identify==2.5.10
    # via pre-commit
iniconfig==1.1.1
    # via pytest
nodeenv==1.7.0
    # via pre-commit
packaging==22.0
    # via pytest
platformdirs==2.6.0
    # via virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.20.0
    # via -r requirements-dev.in
psycopg2-binary==2.9.4
    # via
    #   -c requirements.txt
    #   -r requirements.txt
pytest==7.2.0
    # via
    #   -r requirements-dev.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements-dev.in
python-dateutil==2.8.2
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   faker
pytz==2022.6
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   djangorestframework
pyyaml==6.0
    # via pre-commit
six==1.16.0
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   faker
    #   python-dateutil
sqlparse==0.4.3
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   django
text-unidecode==1.2
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   faker
toml==0.10.2
    # via pre-commit
tomli==2.0.1
    # via pytest
virtualenv==20.17.1
    # via pre-commit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
